fixed the big merg/rebase screw up
